Replace poolboy with pg2
Redis operations are rapid enough that serializing on a single gen_server
as in poolboy quickly becomes a bottleneck. pg2 is backed by ets and provides
a simple random-access pattern across the various redis clients.

Note that this implementation does not easily allow _reducing_ the number
of outstanding clients, since each is booted as a permanent module under
erp_sup; that could be done, however, by rebooting the application as a whole.